Check for reused [localhost]:port
SSH stores the hostkey for each [localhost]:port used in
connecting jump boxes. The issue is that ports are randomly
allocated, so it is possible to attempt to use the same port
with a different on-premise cluster (or clusters with multiple
login nodes). When we try to reuse a port, the Parsl SSHChannel
panics because it thinks there is a hostkey mismatch (i.e.
man-in-the-middle attach).  Solve the problem with ssh-keygen -R
which will remove that specific [localhost]:<port> entry in
~/.ssh/known_hosts in only the PW IDE.